server/pyproject.toml
[tool.poetry]
name = "realtime_graph"
version = "0.1.0"
description = ""
authors = ["Ilya Monakov <imonakov@hostcomm.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.1"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
attrs = "^21.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
